+++ lizmap_server/context/py_qgis_server.py
@@ -1,4 +1,3 @@
-
 from functools import cached_property
 
 from itertools import chain